feat: extract nav menu content
